LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/anb-serialization-expected.txt


PASS "1" becomes "1" 
FAIL "+1" becomes "1" assert_equals: expected "1" but got "+1"
PASS "-1" becomes "-1" 
FAIL "0n + 0" becomes "0" assert_equals: expected "0" but got "0n + 0"
FAIL "0n + 1" becomes "1" assert_equals: expected "1" but got "0n + 1"
FAIL "0n - 1" becomes "-1" assert_equals: expected "-1" but got "0n - 1"
FAIL "1n" becomes "n" assert_equals: expected "n" but got "1n"
FAIL "1n - 0" becomes "n" assert_equals: expected "n" but got "1n - 0"
FAIL "1n + 1" becomes "n+1" assert_equals: expected "n+1" but got "1n + 1"
FAIL "1n - 1" becomes "n-1" assert_equals: expected "n-1" but got "1n - 1"
FAIL "-1n" becomes "-n" assert_equals: expected "-n" but got "-1n"
FAIL "-1n - 0" becomes "-n" assert_equals: expected "-n" but got "-1n - 0"
FAIL "-1n + 1" becomes "-n+1" assert_equals: expected "-n+1" but got "-1n + 1"
FAIL "-1n - 1" becomes "-n-1" assert_equals: expected "-n-1" but got "-1n - 1"
FAIL "+n+1" becomes "n+1" assert_equals: expected "n+1" but got "+n+1"
PASS "-n-1" becomes "-n-1" 
FAIL "n + 0" becomes "n" assert_equals: expected "n" but got "n + 0"
FAIL "n - 0" becomes "n" assert_equals: expected "n" but got "n - 0"
FAIL "2n + 2" becomes "2n+2" assert_equals: expected "2n+2" but got "2n + 2"
FAIL "-2n - 2" becomes "-2n-2" assert_equals: expected "-2n-2" but got "-2n - 2"
